Minor error update

+++ back/src/main/java/com/uade/back/controller/CartController.java
@@ -1,12 +1,19 @@
 package com.uade.back.controller;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-import com.uade.back.service.cart.CartService;
 import com.uade.back.dto.cart.AddItemRequest;
-import com.uade.back.dto.cart.UpdateItemRequest;
 import com.uade.back.dto.cart.CartResponse;
+import com.uade.back.dto.cart.UpdateItemRequest;
+import com.uade.back.service.cart.CartService;
 
 import lombok.RequiredArgsConstructor;
 
@@ -28,12 +35,12 @@ public class CartController {
   }
 
   @PatchMapping("/items/{itemId}")
-  public ResponseEntity<CartResponse> updateItem(@PathVariable Long itemId, @RequestBody UpdateItemRequest request) {
+  public ResponseEntity<CartResponse> updateItem(@PathVariable Integer itemId, @RequestBody UpdateItemRequest request) {
     return ResponseEntity.ok(service.updateItem(itemId, request));
   }
 
   @DeleteMapping("/items/{itemId}")
-  public ResponseEntity<CartResponse> removeItem(@PathVariable Long itemId) {
+  public ResponseEntity<CartResponse> removeItem(@PathVariable Integer itemId) {
     return ResponseEntity.ok(service.removeItem(itemId));
   }
